pyproject.toml
[tool.poetry]
name = "lootmarshal"
version = "0.3.0"
description = ""
authors = ["Leron Gray <daddycocoaman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.55.1"
typer = "^0.2.1"
python-multipart = "^0.0.5"
uvicorn = "^0.11.5"
typer-cli = "^0.0.9"
click-help-colors = "^0.8"
azure-storage-blob = "^12.3.1"
python-dotenv = "^0.13.0"
azure-identity = "^1.3.1"
aiohttp = "^3.6.2"
azure-keyvault-secrets = "^4.1.0"
requests-toolbelt = "^0.9.1"
click-spinner = "^0.1.10"
azure-storage-file = "^2.1.0"
orjson = "^3.1.2"
cryptography = "^2.9.2"
loguru = "^0.5.1"
pypykatz = "^0.3.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
autopep8 = "^1.5.2"
flake8 = "^3.8.2"
mypy = "^0.770"
isort = "^5.0.5"
autoflake = "^1.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
